pyproject.toml
[tool.poetry]
name = "contextvar-repro"
version = "0.1.0"
description = ""
authors = ["Emeka C. Anyanwu <emeka@emeka.ca>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sf-hamilton = "1.87.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
